Updated Type of Player Stat to Double

+++ src/main/java/com/dhl/g05/db/StoredProcedure.java
@@ -251,7 +251,7 @@ public class StoredProcedure {
 		return result;
 	}
 
-	public int savePlayer(int team_id,int position_id,String player_name , int player_is_captain, int age, int skating, int shooting, int checking, int saving)
+	public int savePlayer(int team_id,int position_id,String player_name , int player_is_captain, double age, double skating, double shooting, double checking, double saving)
 	//public int savePlayer(int team_id,int position_id,String player_name , int player_is_captain)
 	{
 		int result=0;
@@ -264,11 +264,11 @@ public class StoredProcedure {
 			stmt.setInt(2,position_id);
 			stmt.setString(3,player_name);
 			stmt.setInt(4,player_is_captain);
-			stmt.setInt(5,age);
-			stmt.setInt(6,skating);
-			stmt.setInt(7,shooting);
-			stmt.setInt(8,checking);
-			stmt.setInt(9,saving);
+			stmt.setDouble(5,age);
+			stmt.setDouble(6,skating);
+			stmt.setDouble(7,shooting);
+			stmt.setDouble(8,checking);
+			stmt.setDouble(9,saving);
 			rs = stmt.executeQuery();
 			while(rs.next())
 			{
@@ -365,7 +365,7 @@ public class StoredProcedure {
 		return result;
 	}
 
-	public int saveFreeAgent(String agent_name, String position_name , String league_name, int age, int skating, int shooting, int checking, int saving)
+	public int saveFreeAgent(String agent_name, String position_name , String league_name, double age, double skating, double shooting, double checking, double saving)
 	//public int saveFreeAgent(String agent_name, String position_name , String league_name)
 	{
 		int result=0;
@@ -377,11 +377,11 @@ public class StoredProcedure {
 			stmt.setString(1, agent_name);
 			stmt.setString(2, position_name);
 			stmt.setString(3, league_name);
-			stmt.setInt(4,age);
-			stmt.setInt(5,skating);
-			stmt.setInt(6,shooting);
-			stmt.setInt(7,checking);
-			stmt.setInt(8,saving);
+			stmt.setDouble(4,age);
+			stmt.setDouble(5,skating);
+			stmt.setDouble(6,shooting);
+			stmt.setDouble(7,checking);
+			stmt.setDouble(8,saving);
 
 			rs = stmt.executeQuery();
 			while(rs.next())
